% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackmagic-package.R
\docType{package}
\name{blackmagic}
\alias{blackmagic}
\alias{blackmagic-package}
\title{Convert XML to JSON}
\description{
Convert XML to JSON. Uses the 'xml-js' 'npm' library
\url{https://www.npmjs.com/package/xml-js} by Yousuf Almarzooqi.
}
\author{
Bob Rudis (bob@rud.is)
}
